+++ packages/eslint-plugin/src/rules/brace-style.ts
@@ -16,6 +16,8 @@ export type MessageIds = InferMessageIdsTypeFromRule<typeof baseRule>;
 export default createRule<Options, MessageIds>({
   name: 'brace-style',
   meta: {
+    deprecated: true,
+    replacedBy: ['@stylistic/brace-style'],
     type: 'layout',
     docs: {
       description: 'Enforce consistent brace style for blocks',