+++ packages/eslint-plugin/src/rules/comma-dangle.ts
@@ -46,6 +46,8 @@ function normalizeOptions(options: Option): NormalizedOptions {
 export default createRule<Options, MessageIds>({
   name: 'comma-dangle',
   meta: {
+    deprecated: true,
+    replacedBy: ['@stylistic/comma-dangle'],
     type: 'layout',
     docs: {
       description: 'Require or disallow trailing commas',