+++ packages/eslint-plugin/src/rules/key-spacing.ts
@@ -27,6 +27,8 @@ const baseSchema = Array.isArray(baseRule.meta.schema)
 export default createRule<Options, MessageIds>({
   name: 'key-spacing',
   meta: {
+    deprecated: true,
+    replacedBy: ['@stylistic/key-spacing'],
     type: 'layout',
     docs: {
       description: