+++ packages/eslint-plugin/src/rules/lines-around-comment.ts
@@ -50,6 +50,8 @@ function getCommentLineNums(comments: TSESTree.Comment[]): number[] {
 export default createRule<Options, MessageIds>({
   name: 'lines-around-comment',
   meta: {
+    deprecated: true,
+    replacedBy: ['@stylistic/lines-around-comment'],
     type: 'layout',
     docs: {
       description: 'Require empty lines around comments',