+++ packages/eslint-plugin/tests/docs.test.ts
@@ -8,13 +8,20 @@ import rules from '../src/rules';
 const docsRoot = path.resolve(__dirname, '../docs/rules');
 const rulesData = Object.entries(rules);
 
-function parseMarkdownFile(filePath: string): marked.TokensList {
-  const file = fs.readFileSync(filePath, 'utf-8');
+interface ParsedMarkdownFile {
+  fullText: string;
+  tokens: marked.TokensList;
+}
 
-  return marked.lexer(file, {
+function parseMarkdownFile(filePath: string): ParsedMarkdownFile {
+  const fullText = fs.readFileSync(filePath, 'utf-8');
+
+  const tokens = marked.lexer(fullText, {
     gfm: true,
     silent: false,
   });
+
+  return { fullText, tokens };
 }
 
 type TokenType = marked.Token['type'];
@@ -46,6 +53,7 @@ describe('Validating rule docs', () => {
     // comments in the files for more information.
     'camelcase.md',
     'no-duplicate-imports.md',
+    'no-parameter-properties.md',
   ]);
   it('All rules must have a corresponding rule doc', () => {
     const files = fs
@@ -63,7 +71,7 @@ describe('Validating rule docs', () => {
 
     describe(`${ruleName}.md`, () => {
       const filePath = path.join(docsRoot, `${ruleName}.md`);
-      const tokens = parseMarkdownFile(filePath);
+      const { fullText, tokens } = parseMarkdownFile(filePath);
 
       test(`${ruleName}.md must start with frontmatter description`, () => {
         expect(tokens[0]).toMatchObject({
@@ -91,31 +99,53 @@ describe('Validating rule docs', () => {
         });
       });
 
-      test(`headers must be title-cased`, () => {
-        // Get all H2 headers objects as the other levels are variable by design.
-        const headers = tokens.filter(tokenIsH2);
+      test(`headings must be title-cased`, () => {
+        // Get all H2 headings objects as the other levels are variable by design.
+        const headings = tokens.filter(tokenIsH2);
 
-        headers.forEach(header =>
-          expect(header.text).toBe(titleCase(header.text)),
+        headings.forEach(heading =>
+          expect(heading.text).toBe(titleCase(heading.text)),
         );
       });
 
+      const requiredHeadings = ['When Not To Use It'];
       const importantHeadings = new Set([
+        ...requiredHeadings,
         'How to Use',
         'Options',
         'Related To',
-        'When Not To Use It',
       ]);
 
       test('important headings must be h2s', () => {
-        const headers = tokens.filter(tokenIsHeading);
+        const headings = tokens.filter(tokenIsHeading);
 
-        for (const header of headers) {
-          if (importantHeadings.has(header.raw.replace(/#/g, '').trim())) {
-            expect(header.depth).toBe(2);
+        for (const heading of headings) {
+          if (importantHeadings.has(heading.raw.replace(/#/g, '').trim())) {
+            expect(heading.depth).toBe(2);
           }
         }
       });
+
+      if (!rules[ruleName as keyof typeof rules].meta.docs?.extendsBaseRule) {
+        test('must include required headings', () => {
+          const headingTexts = new Set(
+            tokens.filter(tokenIsH2).map(token => token.text),
+          );
+
+          for (const requiredHeading of requiredHeadings) {
+            const omissionComment = `<!-- Intentionally Omitted: ${requiredHeading} -->`;
+
+            if (
+              !headingTexts.has(requiredHeading) &&
+              !fullText.includes(omissionComment)
+            ) {
+              throw new Error(
+                `Expected a '${requiredHeading}' heading or comment like ${omissionComment}.`,
+              );
+            }
+          }
+        });
+      }
     });
   }
 });
@@ -130,7 +160,7 @@ describe('Validating rule metadata', () => {
   }
 
   for (const [ruleName, rule] of rulesData) {
-    describe(`${ruleName}`, () => {
+    describe(ruleName, () => {
       it('`name` field in rule must match the filename', () => {
         // validate if rule name is same as url
         // there is no way to access this field but its used only in generation of docs url