+++ packages/typescript-estree/src/parseSettings/createParseSettings.ts
@@ -1,5 +1,5 @@
 import debug from 'debug';
-import type * as ts from 'typescript';
+import * as ts from 'typescript';
 
 import type { TypeScriptProjectService } from '../create-program/createProjectService';
 import { createProjectService } from '../create-program/createProjectService';
@@ -23,6 +23,18 @@ const log = debug(
 let TSCONFIG_MATCH_CACHE: ExpiringCache<string, string> | null;
 let TSSERVER_PROJECT_SERVICE: TypeScriptProjectService | null = null;
 
+// NOTE - we intentionally use "unnecessary" `?.` here because in TS<5.3 this enum doesn't exist
+// This object exists so we can centralize these for tracking and so we don't proliferate these across the file
+// https://github.com/microsoft/TypeScript/issues/56579
+/* eslint-disable @typescript-eslint/no-unnecessary-condition */
+const JSDocParsingMode = {
+  ParseAll: ts.JSDocParsingMode?.ParseAll,
+  ParseNone: ts.JSDocParsingMode?.ParseNone,
+  ParseForTypeErrors: ts.JSDocParsingMode?.ParseForTypeErrors,
+  ParseForTypeInfo: ts.JSDocParsingMode?.ParseForTypeInfo,
+} as const;
+/* eslint-enable @typescript-eslint/no-unnecessary-condition */
+
 export function createParseSettings(
   code: ts.SourceFile | string,
   options: Partial<TSESTreeOptions> = {},
@@ -34,6 +46,22 @@ export function createParseSettings(
       ? options.tsconfigRootDir
       : process.cwd();
   const passedLoggerFn = typeof options.loggerFn === 'function';
+  const jsDocParsingMode = ((): ts.JSDocParsingMode => {
+    switch (options.jsDocParsingMode) {
+      case 'all':
+        return JSDocParsingMode.ParseAll;
+
+      case 'none':
+        return JSDocParsingMode.ParseNone;
+
+      case 'type-info':
+        return JSDocParsingMode.ParseForTypeInfo;
+
+      default:
+        return JSDocParsingMode.ParseAll;
+    }
+  })();
+
   const parseSettings: MutableParseSettings = {
     allowInvalidAST: options.allowInvalidAST === true,
     code,
@@ -56,7 +84,7 @@ export function createParseSettings(
         process.env.TYPESCRIPT_ESLINT_EXPERIMENTAL_TSSERVER !== 'false') ||
       (process.env.TYPESCRIPT_ESLINT_EXPERIMENTAL_TSSERVER === 'true' &&
         options.EXPERIMENTAL_useProjectService !== false)
-        ? (TSSERVER_PROJECT_SERVICE ??= createProjectService())
+        ? (TSSERVER_PROJECT_SERVICE ??= createProjectService(jsDocParsingMode))
         : undefined,
     EXPERIMENTAL_useSourceOfProjectReferenceRedirect:
       options.EXPERIMENTAL_useSourceOfProjectReferenceRedirect === true,
@@ -71,6 +99,7 @@ export function createParseSettings(
         : getFileName(options.jsx),
       tsconfigRootDir,
     ),
+    jsDocParsingMode,
     jsx: options.jsx === true,
     loc: options.loc === true,
     log:
@@ -134,6 +163,17 @@ export function createParseSettings(
       singleRun: parseSettings.singleRun,
       tsconfigRootDir: tsconfigRootDir,
     });
+  }
+
+  // No type-aware linting which means that cross-file (or even same-file) JSDoc is useless
+  // So in this specific case we default to 'none' if no value was provided
+  if (
+    options.jsDocParsingMode == null &&
+    parseSettings.projects.length === 0 &&
+    parseSettings.programs == null &&
+    parseSettings.EXPERIMENTAL_projectService == null
+  ) {
+    parseSettings.jsDocParsingMode = JSDocParsingMode.ParseNone;
   }
 
   warnAboutTSVersion(parseSettings, passedLoggerFn);